account for file icon
When `dashboard-set-file-icons` is non-nil, by default it adds a unicode
file icon and a space before an item on dashboard.

Before this commit, when envoking either `ace-link-dashboard` or
`ace-link-dashboard-remove`, the overlay drawn by `avy` appears on top
of the unicode file icon and makes it unintelligible, regardless of the
value of `avy-style`.

This commit detects unicode file icons when collecting the widgets and
move forword the point past the icon and the padded space. Therefore the
overlay would be drawn correctly on the item.

In addition, this commit also surpresses the warning the
`dashboard-remove-item-under` is not defined.