fix: ace-link-dashboard-remove
Before this commit, `ace-link-dashboard-remove` would also include
widgets user added below the banner. But since they don't belong to any
known section of `dashboard-mode` buffer, a user-error would be thrown.

This commit adds an optional argument `in-section` to
`ace-link-dashboard--collect`. This argument ensures that the widget collected
is within a known section of `dashboard-mode` buffer.

This test is done via the new function
`ace-link-dashboard--in-section-p`.